Material properties: new material properties mechanism
A material has two properties that define its behavior with incoming
light. The diffuse property describe how it reflects light: how we
see it, and the specular property describe how it shines. Each
property has a type ("lambert" for diffuse and "phong" or "blinn"
for specular) that defines the underlaying algorithm and a color.
Blinn and Phong properties also have a power property to describe
the impact of the effect.

Now a material use the following syntax in the scene file:
<material reflection="0.5>
    <diffuse type="lambert" red="0.5" green="0.8" blue="0.2" />
    <specular type="phong" red="1.0" green="0.7" blue="0.4" power="30" />
</material>